requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==22.2.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.5.2
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
